Move device table management to new service MagentaTVDeviceManager
